requirements.txt
celery==5.3.4
Django==4.2.5
djangorestframework==3.14.0
redis==5.0.0
rsa==4.9
